added projectile collision with mobs

+++ src/com/ipsum/Game.java
@@ -157,8 +157,8 @@ public class Game extends Canvas implements Runnable
 		// all rendering here
 
 
-		int xScroll = player.getX() - screen.width / 2;
-		int yScroll = player.getY() - screen.height / 2;
+		int xScroll =(int) player.getX() - screen.width / 2;
+		int yScroll =(int) player.getY() - screen.height / 2;
 
 		level.render(xScroll, yScroll, screen);
